Reuse FacilitatorDashboard for AdminDashboard
- Make Facilitator dashboard as components
